\name{svg-internal}
\alias{svgOpen}
\alias{svgClose}
\alias{svgStartGroup}
\alias{svgEndGroup}
\alias{svgRaster}
\alias{svgRect}
\alias{svgLines}
\alias{svgPath}
\alias{svgPolygon}
\alias{svgText}
\alias{svgCircle}
\alias{svgScript}
\alias{svgAnimate}
\alias{svgStartLink}
\alias{svgEndLink}
\alias{svgStyle}
\title{Internal gridSVG Functions}
\description{
  Some functions you can call, but I'm not documenting (yet).
}
\keyword{internal}
